Flags separated to classic and inverse. Expired Inverse flag added
