+++ tests/database/factories/EntityWithAcceptedFlagFactory.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-$factory->define(\Cog\Flag\Tests\Stubs\Models\EntityWithAcceptedFlag::class, function (\Faker\Generator $faker) {
+$factory->define(\Cog\Flag\Tests\Stubs\Models\Classic\EntityWithAcceptedFlag::class, function (\Faker\Generator $faker) {
     return [
         'name' => $faker->word,
         'is_accepted' => true,
