+++ tests/database/factories/EntityWithExpiredFlagFactory.php
@@ -9,9 +9,9 @@
  * file that was distributed with this source code.
  */
 
-$factory->define(\Cog\Flag\Tests\Stubs\Models\Classic\EntityWithVerifiedFlag::class, function (\Faker\Generator $faker) {
+$factory->define(\Cog\Flag\Tests\Stubs\Models\Inverse\EntityWithExpiredFlag::class, function (\Faker\Generator $faker) {
     return [
         'name' => $faker->word,
-        'is_verified' => true,
+        'is_expired' => false,
     ];
 });
