+++ tests/database/factories/EntityWithPublishedFlagFactory.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-$factory->define(\Cog\Flag\Tests\Stubs\Models\EntityWithPublishedFlag::class, function (\Faker\Generator $faker) {
+$factory->define(\Cog\Flag\Tests\Stubs\Models\Classic\EntityWithPublishedFlag::class, function (\Faker\Generator $faker) {
     return [
         'name' => $faker->word,
         'is_published' => true,
