+++ tests/stubs/Models/Classic/EntityWithActiveFlag.php
@@ -1,0 +1,50 @@
+<?php
+
+/*
+ * This file is part of Laravel Eloquent Flag.
+ *
+ * (c) CyberCog <support@cybercog.su>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Cog\Flag\Tests\Stubs\Models\Classic;
+
+use Cog\Flag\Traits\Classic\HasActiveFlag;
+use Illuminate\Database\Eloquent\Model;
+
+/**
+ * Class EntityWithActiveFlag.
+ *
+ * @package Cog\Flag\Tests\Stubs\Models\Classic
+ */
+class EntityWithActiveFlag extends Model
+{
+    use HasActiveFlag;
+
+    /**
+     * The table associated with the model.
+     *
+     * @var string
+     */
+    protected $table = 'entity_with_active_flag';
+
+    /**
+     * The attributes that are mass assignable.
+     *
+     * @var array
+     */
+    protected $fillable = [
+        'name',
+    ];
+
+    /**
+     * The attributes that should be cast to native types.
+     *
+     * @var array
+     */
+    protected $casts = [
+        'is_active' => 'bool',
+    ];
+}
